MdeModulePkg PeiCore: Update debug message to print FV handle correctly
BfvHeader->FvLength is UINT64. Now, it prints with %x. It will cause the
late FvHandle to be as zero. So, its type is converted to UINT32.

Contributed-under: TianoCore Contribution Agreement 1.1
Cc: Star Zeng <star.zeng@intel.com>

+++ MdeModulePkg/Core/Pei/FwVol/FwVol.c
@@ -514,7 +514,7 @@ PeiInitializeFv (
     "The %dth FV start address is 0x%11p, size is 0x%08x, handle is 0x%p\n", 
     (UINT32) PrivateData->FvCount, 
     (VOID *) BfvHeader, 
-    BfvHeader->FvLength,
+    (UINT32) BfvHeader->FvLength,
     FvHandle
     ));    
   PrivateData->FvCount ++;
